+++ packages/atomic/src/components/search/atomic-refine-modal/atomic-refine-modal.tsx
@@ -23,6 +23,7 @@ import CloseIcon from 'coveo-styleguide/resources/icons/svg/close.svg';
 import {getFacetElements, SortDropdownOption} from '../../../utils/store';
 import SortIcon from '../../../images/sort.svg';
 import {Button} from '../../common/button';
+import {BaseFacetElement} from '../facets/facet-common';
 
 /**
  * The `atomic-refine-modal` is automatically created as a child of the `atomic-search-interface` when the `atomic-refine-toggle` is initialized.
@@ -111,8 +112,8 @@ export class AtomicRefineModal implements InitializableComponent {
       .map((f) => f.payload);
 
     sortedFacetsElements.forEach((facetElement) => {
-      const clone = facetElement.cloneNode(true) as HTMLElement;
-      clone.setAttribute('is-collapsed', 'true');
+      const clone = facetElement.cloneNode(true) as BaseFacetElement;
+      clone.isCollapsed = true;
       divSlot.append(clone);
     });
 
@@ -248,6 +249,7 @@ export class AtomicRefineModal implements InitializableComponent {
         fullscreen
         isOpen={this.isOpen}
         source={this.openButton}
+        container={this.host}
         close={() => (this.isOpen = false)}
         onAnimationEnded={() => this.onAnimationEnded()}
         exportparts="container,header,header-wrapper,header-ruler,body,body-wrapper,footer,footer-wrapper,footer-wrapper"